pyproject.toml
[project]
name = "hr_bot"
version = "7.1"
description = "An HR assistant powered by AWS Bedrock Nova Lite v1 and CrewAI with ETag-based S3 smart caching, role-based document access, and enhanced RAG filtering"
authors = [{ name = "Saish Shinde", email = "saish.shinde15@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    # Core CrewAI
    "crewai[tools]>=0.130.0,<1.0.0",
    # Langchain & AI
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    # AWS Bedrock support
    "boto3>=1.28.0",
    "langchain-aws>=0.1.0",
    # Vector Store & Search
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",
    # Document Processing
    "unstructured[docx]>=0.12.0",
    "python-docx>=1.1.0",
    "docx2txt>=0.8",
    # HTTP & API
    "httpx>=0.27.0",
    # Caching & Performance
    "diskcache>=5.6.3",
    # Utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    # Local Embeddings (HuggingFace)
    "langchain-huggingface>=0.3.0",
    "sentence-transformers==2.7.0",
    "torch==2.6.0",
    "pytest>=8.2.0",
    # Evaluation & hallucination metrics
    "ragas>=0.1.20",
    "datasets>=2.19.0",
    "evaluate>=0.4.1",
    # UI
    # Streamlit is retained as a legacy/rollback UI. Move to optional deps below.
    # FastAPI & Production API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pyjwt[crypto]>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.6",
    "authlib>=1.6.5",
    "requests>=2.32.5",
    "starlette>=0.49.3",
    "itsdangerous>=2.2.0",
    "chainlit>=2.9.0",
]

[project.scripts]
hr_bot = "hr_bot.main:run"
hr_bot_interactive = "hr_bot.main:interactive"
hr_bot_setup = "hr_bot.main:setup"
run_crew = "hr_bot.main:run_crew"
train = "hr_bot.main:train"
replay = "hr_bot.main:replay"
test = "hr_bot.main:test"
generate_eval = "hr_bot.eval.generate_dataset:main"
eval = "hr_bot.eval.run_eval:main"
ragas_eval = "hr_bot.eval.ragas_eval:main"
ragas_prod_eval = "hr_bot.eval.ragas_production_eval:main"
api_server = "hr_bot.api.main:main"
chainlit_app = "chainlit:cli"

[project.optional-dependencies]
legacy = ["streamlit>=1.38.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
